requirements.txt
certifi==2020.12.5
click==7.1.2
fastapi==0.78.0
h11==0.11.0
httptools==0.1.1
mangum==0.10.0
pydantic==1.9.1
python-dotenv==0.15.0
python-json-logger==2.0.1
PyYAML==5.3.1
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.14.0
watchgod==0.6
websockets==8.1
hgtk==0.1.3
meilisearch==0.18.3
idna==3.3
requests==2.28.1
requests-toolbelt==0.9.1
requestsexceptions==1.4.0
charset-normalizer==2.1.0
urllib3==1.26.9
fastapi-simple-security==1.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.39
fastapi-sqlalchemy==0.2.1
six==1.16.0
passlib[bcrypt]
python-jose[cryptography]
passlib
python-jose
python-multipart
bcrypt
boto3